Remove interchange out of the python API
