+++ py-polars/src/polars/__init__.py
@@ -60,6 +60,7 @@ from polars._utils.polars_version import get_polars_version as _get_polars_versi
 # TODO: remove need for importing wrap utils at top level
 from polars._utils.wrap import wrap_df, wrap_s  # noqa: F401
 from polars.catalog.unity import Catalog
+from polars.compat_level import CompatLevel
 from polars.config import Config
 from polars.convert import (
     from_arrow,
@@ -211,7 +212,6 @@ from polars.functions import (
     when,
     zeros,
 )
-from polars.interchange import CompatLevel
 from polars.io import (
     FileProviderArgs,
     PartitionBy,
