--- py-polars/src/polars/interchange/__init__.py
@@ -1,10 +1,0 @@
-"""
-Module containing the implementation of the Python dataframe interchange protocol.
-
-Details on the protocol:
-https://data-apis.org/dataframe-protocol/latest/index.html
-"""
-
-from polars.interchange.protocol import CompatLevel
-
-__all__ = ["CompatLevel"]
