--- py-polars/tests/unit/interchange/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for dataframe interchange protocol support."""
